feat(settings): 在设置底部常驻显示构建 / SW 版本
正式版（main/master）会树摇掉右下角 BuildBadge，导致线上拿不到
构建分支 / commit / SW 版本，用户报障时缺少版本上下文。

- 抽出 querySwVersion 到 utils/swVersion.ts，BuildBadge 与新组件共用
- 新增 components/settings/VersionInfo.tsx：设置页底部常驻脚注，
  在所有构建（含 prod）显示 APP 版本 + build branch@commit + sw 版本，
  样式为居中两枚低调 chip，不喧宾夺主
- Settings 底部替换原硬编码 v2.2 文案为 <VersionInfo />

构建全局由 vite define 始终注入，prod 也有值，无需 dev 条件判断。
